修改方法
