+++ src/main/java/com/fc/test/service/SysAreaService.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
-import com.fc.test.common.support.Convert;
+import com.fc.test.common.support.ConvertUtil;
 import com.fc.test.mapper.auto.SysAreaMapper;
 import com.fc.test.model.auto.SysArea;
 import com.fc.test.model.auto.SysAreaExample;
@@ -48,7 +48,7 @@ public class SysAreaService {
 
 	
 	public int deleteByPrimaryKey(String ids) {
-		Integer[] integers = Convert.toIntArray(",", ids);
+		Integer[] integers = ConvertUtil.toIntArray(",", ids);
 		List<Integer> stringB = Arrays.asList(integers);
 		SysAreaExample example=new SysAreaExample();
 		example.createCriteria().andIdIn(stringB);
