+++ src/main/java/com/fc/test/service/SysNoticeService.java
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.fc.test.common.base.BaseService;
-import com.fc.test.common.support.Convert;
+import com.fc.test.common.support.ConvertUtil;
 import com.fc.test.mapper.auto.SysNoticeMapper;
 import com.fc.test.mapper.auto.SysNoticeUserMapper;
 import com.fc.test.model.auto.SysNotice;
@@ -107,7 +107,7 @@ public class SysNoticeService implements BaseService<SysNotice, SysNoticeExample
 	 
 	@Override
 	public int deleteByPrimaryKey(String ids) {
-		List<String> lista=Convert.toListStrArray(ids);
+		List<String> lista=ConvertUtil.toListStrArray(ids);
 		SysNoticeExample example=new SysNoticeExample();
 		example.createCriteria().andIdIn(lista);
 		return sysNoticeMapper.deleteByExample(example);
