+++ src/main/java/com/fc/test/service/SysNoticeUserService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.fc.test.common.base.BaseService;
-import com.fc.test.common.support.Convert;
+import com.fc.test.common.support.ConvertUtil;
 import com.fc.test.mapper.auto.SysNoticeUserMapper;
 import com.fc.test.model.auto.SysNoticeUser;
 import com.fc.test.model.auto.SysNoticeUserExample;
@@ -47,7 +47,7 @@ public class SysNoticeUserService implements BaseService<SysNoticeUser, SysNotic
 
 	@Override
 	public int deleteByPrimaryKey(String ids) {
-		List<String> lista=Convert.toListStrArray(ids);
+		List<String> lista=ConvertUtil.toListStrArray(ids);
 		SysNoticeUserExample example=new SysNoticeUserExample();
 		example.createCriteria().andIdIn(lista);
 		return sysNoticeUserMapper.deleteByExample(example);
