+++ src/main/java/com/fc/test/service/SysOperLogService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.fc.test.common.base.BaseService;
-import com.fc.test.common.support.Convert;
+import com.fc.test.common.support.ConvertUtil;
 import com.fc.test.mapper.auto.TsysOperLogMapper;
 import com.fc.test.model.auto.TsysOperLog;
 import com.fc.test.model.auto.TsysOperLogExample;
@@ -42,7 +42,7 @@ public class SysOperLogService implements BaseService<TsysOperLog, TsysOperLogEx
 	
 	@Override
 	public int deleteByPrimaryKey(String ids) {
-		List<String> lista=Convert.toListStrArray(ids);
+		List<String> lista=ConvertUtil.toListStrArray(ids);
 		TsysOperLogExample example=new TsysOperLogExample();
 		example.createCriteria().andIdIn(lista);
 		return tsysOperLogMapper.deleteByExample(example);
