+++ src/main/resources/auto_code/service/EntityService.java.vm
@@ -8,7 +8,7 @@ import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import cn.hutool.core.util.StrUtil;
 import ${globalConfig.parentPack}.common.base.BaseService;
-import ${globalConfig.parentPack}.common.support.Convert;
+import ${globalConfig.parentPack}.common.support.ConvertUtil;
 import ${globalConfig.parentPack}.mapper.auto.${TsysTables.tableModel}Mapper;
 import ${globalConfig.parentPack}.model.auto.${TsysTables.tableModel};
 import ${globalConfig.parentPack}.model.auto.${TsysTables.tableModel}Example;
@@ -68,14 +68,14 @@ public class ${TsysTables.tableModel}Service implements BaseService<${TsysTables
 	public int deleteByPrimaryKey(String ids) {
 		#if(${pk_attrType} == 'String')
 		
-			List<String> lista=Convert.toListStrArray(ids);
+			List<String> lista=ConvertUtil.toListStrArray(ids);
 			${TsysTables.tableModel}Example example=new ${TsysTables.tableModel}Example();
 			example.createCriteria().and${pkfield}In(lista);
 			return ${TsysTables.tableModel_a}Mapper.deleteByExample(example);
 			
 		#elseif(${pk_attrType} == 'Long')
 		
-			Long[] integers = Convert.toLongArray(",", ids);
+			Long[] integers = ConvertUtil.toLongArray(",", ids);
 			List<Long> stringB = Arrays.asList(integers);
 			${TsysTables.tableModel}Example example=new ${TsysTables.tableModel}Example();
 			example.createCriteria().and${pkfield}In(stringB);
@@ -83,7 +83,7 @@ public class ${TsysTables.tableModel}Service implements BaseService<${TsysTables
 		
 		#else
 			
-			Integer[] integers = Convert.toIntArray(",", ids);
+			Integer[] integers = ConvertUtil.toIntArray(",", ids);
 			List<Integer> stringB = Arrays.asList(integers);
 			${TsysTables.tableModel}Example example=new ${TsysTables.tableModel}Example();
 			example.createCriteria().and${pkfield}In(stringB);
